ajout et suppression live

+++ controllers/books.js
@@ -1,4 +1,5 @@
 const Book = require('../models/Book');
+const fs = require('fs');
 
 exports.getAllBooks = (req, res, next) => {
     Book.find()
@@ -6,30 +7,64 @@ exports.getAllBooks = (req, res, next) => {
         .catch(error => res.status(400).json({ error }));
 };
 
+exports.createBook = (req, res, next) => {
+
+    const bookObject = JSON.parse(req.body.book);
+    delete bookObject._id;
+    delete bookObject._userId;
+    const book = new Book({ 
+        ...bookObject,
+        userId: req.auth.userId,
+        imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+     });
+
+     book.save()
+        .then(() => {res.status(201).json({ message: 'Livre enregistré ! '})})
+        .catch(error => res.status(400).json({ error }));
+};
+
+
 exports.getOneBook = (req, res, next) => {
     Book.findOne({ _id: req.params.id })
         .then(book => res.status(201).json(book))
         .catch(error => res.status(404).json({ error }));
 }
 
-exports.createBook = (req, res, next) => {
-    delete req.body._id;
-    const book = new Book({
-        ...req.body
-    });
-    book.save()
-        .then(() => res.status(201).json[{ message : 'Ajout réussi'}])
+exports.modifyBook = (req, res, next) => {
+    const bookObject = req.file ? {
+        ...JSON.parse(req.body.book),
+        imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
+    } :  { ...req.body };
+
+    delete bookObject._userId;
+    Book.findOne({_id: req.params.id})
+        .then((book) => {
+            if (book.userId != req.auth.userId) {
+                res.status(401).json({ message: 'Non-authorisé' });
+            } else {
+                Book.updateOne({_id: req.params.id}, { ...bookObject, _id: req.params.id })
+                .then(() => res.status(200).json({ message:'Livre modifié !'}))
+                .catch(error => res.status(401).json({ error }));
+            }
+        })
         .catch(error => res.status(400).json({ error }));
+
 };
 
-exports.modifyBook = (req, res, next) => {
-    Book.updateOne({ _id: req.params.id }, { ...req.body, _id: req.params.id })
-        .then(() => res.status(200).json({ message: 'Livre modifié'}))
-        .catch(error => res.status(400).json({ error }));
-}
-
 exports.deleteBook = (req, res, next) => {
-    Book.deleteOne({ _id: req.params.id })
-    .then(() => res.status(200).json[{ message : 'Livre supprimé'}])
-    .catch(error => res.status(400).json({ error }));
-}
+    Book.findOne({_id:req.params.id})
+        .then(book => {
+            if (book.userId != req.auth.userId) {
+                res.status(401).json({  message:'Non-authorisé'})
+            } else  {
+                const filename = book.imageUrl.split('/images/')[1];
+                fs.unlink(`images/${filename}`, () => {
+                    Book.deleteOne({_id: req.params.id})
+                        .then(() => { res.status(200).json({ message:'Livre supprimé !' })})
+                        .catch(error => res.status(400).json({ error }));
+                });
+            }
+        })
+        .catch(error => { res.status(500).json({ error })});
+
+};
